feat: add animation for cases title
refactor: add new module prizes-svg-animation.js, add wrapper functions

fix: the numbers in the title change in ascending order
