+++ source/scss/blocks/prizes.scss
@@ -182,6 +182,14 @@
       margin-left: 5rem;
     }
 
+    .prizes__desc b {
+      width: 6rem;
+
+      @media (min-width: $stop-scaling) {
+        width: 60px;
+      }
+    }
+
     .prizes__icon {
       @media (min-width: $tablet) and (orientation: landscape) {
         position: relative;
